+++ ChangeLog.md
@@ -3,7 +3,7 @@
 ## 0.1.3.0  -- 2020-09-19
 
 * Add indexing funtion to `Z.Data.Vector` and `Z.Data.Text`.
-* Rename `allocMutableByteArrayUnsafe` to `allocBytesUnsafe`, add `allocBytesSafe`.
+* Add `peekMBA` and `pokeMBA` to `Z.Foreign`.
 
 ## 0.1.2.0  -- 2020-09-19